feat: Add a basic diesel::Instrumentation impl
Adds a very basic implementation for using `tracing` with diesel's
`Instrumentation` trait. This implementation just maps `diesel` events
directly to `tracing` without much extra information provided beyond
that in the events themselves.

In order to bring this in line with the rest of the crate I think there
are a few more additions that are required before a release:

- Saving the url used during connection to be included in other events.
- Using opentelemetry semantic conventions for event values.
- Usage documentation and recommendations.